chore: refactor project folders to standardise
