+++ SmartHomeManager/SmartHomeManager.Domain/SmartHomeManager.Domain.csproj
@@ -6,4 +6,8 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="RoomDomain\Services\" />
+  </ItemGroup>
+
 </Project>
